requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
Django==4.2.10
django-appconf==1.0.6
django-bootstrap-v5==1.0.11
django-compressor==4.4
django-multiupload==0.6.1
gunicorn==21.2.0
packaging==24.0
pillow==10.2.0
rcssmin==1.1.1
rjsmin==1.2.1
soupsieve==2.5
sqlparse==0.4.4
whitenoise==6.6.0
qrcode==7.4.2
python-barcode==0.15.1
